refactor: split input-audio abstraction from pacer implementation
Extract AudioInputSender (abstract base), AudioInputHost (Protocol), and
DirectInput into a new audio_input_sender.py. audio_input_pacer.py now
contains only the pacer-specific code (AudioInputPacingConfig and
AudioInputPacer, which inherits from DirectInput).

Realtime now types its _audio_input_sender field as the AudioInputSender
abstraction rather than the concrete DirectInput, which honestly conveys
that the field holds any sender, not the direct one specifically.
Rename the field from _input_audio to _audio_input_sender to read
naturally with .send()/.clear()/.close() and to match the AudioInput*
naming used by the surrounding classes.
