+++ agents-core/vision_agents/core/llm/realtime.py
@@ -18,8 +18,10 @@ from vision_agents.core.llm.events import (
 from vision_agents.core.utils.audio_input_pacer import (
     AudioInputPacer,
     AudioInputPacingConfig,
+)
+from vision_agents.core.utils.audio_input_sender import (
+    AudioInputSender,
     DirectInput,
-    InputAudioStrategy,
 )
 from vision_agents.core.utils.stream import Stream
 
@@ -110,7 +112,7 @@ class Realtime(OmniLLM):
         # Store current participant for user speech transcription events
         self._current_participant: Participant | None = None
 
-        self._input_audio: InputAudioStrategy = (
+        self._audio_input_sender: AudioInputSender = (
             AudioInputPacer(
                 self,
                 input_audio_pacing,
@@ -170,7 +172,7 @@ class Realtime(OmniLLM):
         """Increment epoch so stale audio output events are discarded."""
         self._epoch += 1
         self._current_participant = None
-        self._input_audio.clear()
+        self._audio_input_sender.clear()
         self._output.clear()
 
     def _run_tool_in_background(self, coro: Coroutine[None, None, None]) -> None:
@@ -193,10 +195,10 @@ class Realtime(OmniLLM):
 
     async def process_audio(self, pcm: PcmData, participant: Participant) -> None:
         self._current_participant = participant
-        await self._input_audio.send(pcm, participant)
+        await self._audio_input_sender.send(pcm, participant)
 
     async def _close_input_audio(self) -> None:
-        await self._input_audio.close()
+        await self._audio_input_sender.close()
 
     async def stop_watching_video_track(self) -> None:
         """Optionally overridden by providers that support video input."""
@@ -221,7 +223,7 @@ class Realtime(OmniLLM):
     def _on_disconnected(self, reason: str | None = None, clean: bool = True):
         """Mark the session disconnected and emit RealtimeDisconnectedEvent."""
         self.connected = False
-        self._input_audio.clear()
+        self._audio_input_sender.clear()
         self.events.send(
             RealtimeDisconnectedEvent(
                 plugin_name=self.provider_name,
