+++ plugins/gemini/tests/test_gemini_realtime.py
@@ -28,7 +28,8 @@ from vision_agents.core.llm.realtime import (
     RealtimeUserSpeechStarted,
     RealtimeUserTranscript,
 )
-from vision_agents.core.utils.audio_input_pacer import AudioInputPacer, DirectInput
+from vision_agents.core.utils.audio_input_pacer import AudioInputPacer
+from vision_agents.core.utils.audio_input_sender import DirectInput
 from vision_agents.plugins.gemini import Realtime
 from vision_agents.plugins.gemini.gemini_realtime import (
     DEFAULT_MODEL,
@@ -77,11 +78,11 @@ class TestGeminiRealtimeInputPacing:
 
         assert DEFAULT_MODEL == "gemini-3.1-flash-live-preview"
         assert LIVE_TRANSLATE_MODEL == "gemini-3.5-live-translate-preview"
-        assert isinstance(default_rt._input_audio, DirectInput)
-        assert isinstance(translate_rt._input_audio, AudioInputPacer)
-        assert translate_rt._input_audio.config.silence_when_empty
-        assert translate_rt._input_audio.config.startup_buffer_ms == 500
-        assert isinstance(translate_opt_out._input_audio, DirectInput)
+        assert type(default_rt._audio_input_sender) is DirectInput
+        assert isinstance(translate_rt._audio_input_sender, AudioInputPacer)
+        assert translate_rt._audio_input_sender.config.silence_when_empty
+        assert translate_rt._audio_input_sender.config.startup_buffer_ms == 500
+        assert type(translate_opt_out._audio_input_sender) is DirectInput
 
 
 class TestGeminiRealtimeProcessEvents:
